feat: live activity stream, full UI audit, clean repo
Dashboard
- Wire Live Activity Stream on / to real commit data from /ui/bootstrap
  (new activity_stream key — author, branch, repo, short SHA, status badge)
- Replace all hardcoded placeholder values across index.html, queue.html,
  scheduler.html, simulation.html with "—" initialized from JS on first poll
- Scheduler kanban: clear static placeholder cards; add Completed column
  wired to real completed pipeline runs; Decision Log now renders live
  data (running → assigned, queued → awaiting, completed → done)
- Simulation live log: clear static rows; add 10s polling interval so
  the page refreshes independently of the dashboard guard
- Webhooks: remove hardcoded secret and fake ngrok URL; add data-ui attrs
- Empty-state rows for queue table, webhook events table, all kanban columns
- Worker node count text now live ("X of Y nodes active")

Backend
- /ui/scheduler: add completed key — queries recently completed runs
- /ui/bootstrap: add activity_stream list (latest 15 pipeline runs)
- Fix SystemMetrics.memory_total_bytes Integer → BigInteger (int32 overflow
  on machines with >2 GB RAM caused 500 on /ui/metrics/live)
- Add _repo_short_name and _clean_trigger helpers to ui.py

Repo hygiene
- git rm --cached frontend/app/node_modules/ (thousands of files removed)
- git rm --cached celerybeat-schedule.{bak,dat,dir} bootstrap.json
- .gitignore: add node_modules/, celerybeat-schedule.*, bootstrap.json, *.db

README
- Add Dashboard section describing all 7 pages
- Add full API reference table including /ui/* endpoints with poll intervals
- Add NGROK_URL to env var table
- Fix port 8001 → 8000 throughout
- Fix typo "Architectur  e"
- Fix MD060 table separator spacing and MD040 bare code fences

+++ app/models.py
@@ -4,7 +4,7 @@ SQLAlchemy ORM models — core tables.
 
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, DateTime, Integer, String, Text, Float
+from sqlalchemy import BigInteger, Boolean, DateTime, Integer, String, Text, Float
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
@@ -59,8 +59,8 @@ class SystemMetrics(Base):
     id:                 Mapped[int]      = mapped_column(Integer, primary_key=True, autoincrement=True)
     timestamp:          Mapped[datetime] = mapped_column(DateTime(timezone=True), default=_utcnow)
     uptime_seconds:     Mapped[int]      = mapped_column(Integer, nullable=False)
-    memory_used_bytes:  Mapped[int]      = mapped_column(Integer, nullable=False)
-    memory_total_bytes: Mapped[int]      = mapped_column(Integer, nullable=False)
+    memory_used_bytes:  Mapped[int]      = mapped_column(BigInteger, nullable=False)
+    memory_total_bytes: Mapped[int]      = mapped_column(BigInteger, nullable=False)
     cpu_percent:        Mapped[float]    = mapped_column(Float, nullable=False)
     queue_total:        Mapped[int]      = mapped_column(Integer, nullable=False)
     busy_workers:       Mapped[int]      = mapped_column(Integer, nullable=False)
